refactor: add index.ts to the reusable comps. folder
Add index.ts file to the reusable components folder and make specific
type alias for that file inside of tsconfig. Remove needless code
leftovers throughout whole project(e.g. commented code or code which not in use).
Replace old paths to the components from the reusable folder
to be up to date with the new convention.